# Conflicts:
#	CookieEngine/include/Resources/Map.hpp
#	CookieEngine/src/ECS/Coordinator.cpp
#	CookieEngine/src/Editor.cpp
#	CookieEngine/src/Resources/Map.cpp

+++ CookieEngine/deps/include/reactphysics3d/collision/narrowphase/NarrowPhaseInfoBatch.h
@@ -126,7 +126,7 @@ struct NarrowPhaseInfoBatch {
 
 /// Return the number of objects in the batch
 inline uint NarrowPhaseInfoBatch::getNbObjects() const {
-    return overlappingPairIds.size();
+    return (uint)overlappingPairIds.size();
 }
 
 }
